In the "Stories" module, I greatly improved the definition of story chapters. I created a new dictionary of chapters, to contain the information for each chapter. Each chapter dictionary has its number, titles in supported languages, and dates of writing, revision, and translation. In its "Write" sub-class, I added information explaining the writing steps, to better inform the user. Such as the opening of the chapter files, server, story website, and music player. I moved the server definition and opening methods to the root class of the "PHP" module, to make it easier to use. I added the ability to pause the writing session, subtracting the pause time from the total duration of the session, with the new "Pause_Writing" method. I also added the ability to postpone the writing session. Just registering the writing time, not writing on Diary Slim, and not asking if the user has finished writing the entire chapter. The sub-class now creates or updates the chapter dictionary, also registering its revision or translation date, not just the writing date. It does this using the new "Update_Chapter_Dictionary" method, the method also updates the chapter titles inside its dictionary.

+++ Modules/Utility/System/__init__.py
@@ -84,7 +84,8 @@ class System():
 		import psutil
 
 		for process in (process for process in psutil.process_iter() if program.split("\\")[program.count("\\")] in process.name()):
-			process.kill()
+			if self.switches["testing"] == False:
+				process.kill()
 
 	def Open_Link(self, link, verbose = True):
 		import webbrowser
